requirements.txt
requests==2.32.3
pyyaml==6.0.2
python-dotenv==1.0.1
gspread
oauth2client
plotly
pandas
flask
